# bff/openapi.yaml
...
openapi: 3.1.0
info:
  title: Miaoyomi API
  version: 0.19.0
  description: "Everything the web app does, it does over this API. Session JWTs are for the browser; scripts
    use a personal API token (Profile -> Security -> API tokens), sent as `Authorization: Bearer uy_...`.
    Scopes: read (every token), write (anything that changes data), admin (the /api/admin surface, which also
    needs an admin account). /img/* takes the yomi_img cookie or the OPDS token over HTTP Basic; /opds/* takes
    the OPDS token over HTTP Basic; neither accepts API tokens. List endpoints answer { content: [...] };
    errors answer a non-2xx status with { error, message }. The prose guide with worked examples is
    docs/api.md."
servers:
  - url: /
    description: This server
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: A session JWT from POST /auth/login, or a personal API token (uy_...). Admin routes need the
        admin scope and an admin account.
    opdsBasic:
      type: http
      scheme: basic
      description: HTTP Basic where the password is the OPDS token from POST /api/opds/token.
    imgCookie:
      type: apiKey
      in: cookie
      name: yomi_img
      description: Set at login so <img> tags can authenticate.
...
paths:
...
  /opds:
    get:
      tags:
        - opds
      summary: Root navigation feed.
      operationId: getOpds
      security:
        - opdsBasic: []
      responses:
        "200":
          description: OK
        "401":
          description: Not authenticated
  /opds/opensearch.xml:
    get:
      tags:
        - opds
      summary: OpenSearch descriptor.
      operationId: getOpdsOpensearchXml
      security:
        - opdsBasic: []
      responses:
        "200":
          description: OK
        "401":
          description: Not authenticated
  /opds/series:
    get:
      tags:
        - opds
      summary: Series listing with facets (sort, library, genre, status), paged sixty at a time.
      operationId: getOpdsSeries
      parameters:
        - name: sort
          in: query
          required: false
          schema:
            type: string
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: library
          in: query
          required: false
          schema:
            type: string
        - name: genre
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
      security:
        - opdsBasic: []
      responses:
        "200":
          description: OK
        "401":
          description: Not authenticated
  /opds/search:
    get:
      tags:
        - opds
      summary: The same feed as /opds/series, reached from a reader's search box.
      operationId: getOpdsSearch
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: sort
          in: query
          required: false
          schema:
            type: string
        - name: library
          in: query
          required: false
          schema:
            type: string
        - name: genre
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
      security:
        - opdsBasic: []
      responses:
        "200":
          description: OK
        "401":
          description: Not authenticated
  /opds/series/{id}:
    get:
      tags:
        - opds
      summary: "A series' chapters: a CBZ acquisition link and an OPDS-PSE stream link per chapter."
      operationId: getOpdsSeriesId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - opdsBasic: []
      responses:
        "200":
          description: OK
        "401":
          description: Not authenticated
        "404":
          description: Not found, or not visible to this account
  /opds/book/{id}/page/{n}:
    get:
      tags:
        - opds
      summary: One page for OPDS-PSE readers. ZERO-based; ?maxWidth (64-2000) returns a JPEG no wider than asked.
      operationId: getOpdsBookIdPageN
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: n
          in: path
          required: true
          schema:
            type: string
        - name: maxWidth
          in: query
          required: false
          schema:
            type: integer
      security:
        - opdsBasic: []
      responses:
        "200":
          description: OK
        "400":
          description: Malformed body or parameter
        "401":
          description: Not authenticated
        "404":
          description: Not found, or not visible to this account
  /opds/book/{id}/file:
    get:
      tags:
        - opds
      summary: Download a chapter as CBZ (zipped on the fly for a loose-image folder).
      operationId: getOpdsBookIdFile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - opdsBasic: []
      responses:
        "200":
          description: OK
        "401":
          description: Not authenticated
        "404":
          description: Not found, or not visible to this account
...